monet: theme: Retarget colors for fixed sRGB conversion
The color targets make much more sense now. Lightness is nearly
perfectly linear, except a sharp decline to black at shade 1000. N-2's
chroma is almost exactly half of N-1's. A-1's chroma is an upside-down V
shape that peaks at shade 500, but we can approximate it with C=0.12 for
now. A-2 and A-3 have constant chroma values.