monet: colors: Add Jzazbz color space with polar representation
From the creator of Oklab:
"Compared with [the CAM16 family], Oklab is missing two major things (by
design): adaptation to viewing conditions and chroma compression. Both
should make Oklab perform worse for color differences (but makes it much
simpler and better for blending colors)" [1]

Wikipedia suggests CAM16-UCS and Jzazbz for color difference with
Euclidean distance [2]. CAM16 is much harder to implement, so let's try
Jzazbz for it first. We'll stick with Oklab for color scheme generation
because it works well for that.

[1] https://twitter.com/bjornornorn/status/1354532019536068610
[2] https://en.wikipedia.org/wiki/Color_difference