ripple: Initial implementation of Fluent-inspired ripple
This is my custom ripple animation inspired by Microsoft's Fluent Design
reveal animation. It renders a wave (ring) with that progressively gets
larger with a smaller blur radius. Blur is approximated with the inverse
square root sigmoid function, and the final output is dithered with
triangular white noise to mitigate banding.