monet: theme: Search for L to target CIELAB L* reference
Ideally, uniform color spaces would have perfectly uniform lightness as
chroma and hue changes, but unfortunately that's not the case. In
particular, colors at the extreme ends of lightness (white and black;
especially shade ~10) diverge from targets significantly.

Searching for the optimal Oklab L value to match the Material You
targets, which specify lightness in terms of CIELAB L*, nearly
eliminates the problem. A binary search is used for speed and accuracy,
similar to what Google does for their CAM16-based implementation.

This could be more efficient in terms of calculations and garbage
creation, but it's fast enough for now.