monet: theme: Preserve chroma ratios when clamping chroma
We clamp chroma instead of setting it to a constant so that low-chroma
colors like brown don't become orange/red. However, clamping chroma on a
per-swatch basis breaks chroma ratios -- e.g. A-1 becomes the same as
A-2.

Scale chroma relative to A-1 / N-1 (depending on the swatch) to fix the
issue while still allowing input colors to express chroma.